do not fail on shallow clones

+++ modules/assets.ts
@@ -4,6 +4,7 @@ import { defineNuxtModule } from '@nuxt/kit'
 export default defineNuxtModule({
    setup(_, nuxt) {
       nuxt.hook('build:before', () => {
+         if (!existsSync('assets/emblem')) return
          if (!existsSync('public')) mkdirSync('public')
          copyFileSync('assets/emblem/favicon.ico', 'public/favicon.ico')
       })
